Handle modified link in non-array property
e.g. [ "@graph", 0, "descriptionLastModifier", "@id" ]
"https://libris.kb.se/library/S" -> "https://libris.kb.se/library/Utb1"

It used to crash because r.val isn't a string because of the special
handling of `@id` in `added/removedPathsInCurrentVersion()`

Uncaught (in promise) TypeError: r.val.concat is not a function
    _callee$/< version-history.vue:189

With this quick fix the new value is shown and the field highlighted as
changed. A better solution woud show both the old and new link.