Fix Title separator in card headings
Separate token components with interpunct (•) instead of space.

Using chips as card headings caused e.g. mainTitle and subtitle to
be displayed with only a space as separator since the nested Title
object is rendered as a token.

Rendering of lenses with multiple levels of data needs to be revisited.