Batch saving/linking of extracted items
The creation of extracted items (and linking to the newly extracted item) is now done first when the user clicks save on an instance (instead of the saving the extracted directly after the user chooses to create a link, which often could result in mistakes and unused works).

The keys to be extracted are batched in the vuex store (`keysToExtractOnSave`), enabling saving of the extracted data iteratively before saving the main instance.