Remove async loading of field and entity-summary
Reverts to using globally registered components as defineAsyncComponent caused unacceptable initial loading times.  Old local imports of the field and entity-summary are removed to normalize the usage of the components.

The components needs to be either globally registered or imported async to resolve issues with circular dependencies (where a component relies on another component, which in turn relies on the first component).