Feature/lws 62 more facets (#1011)
Some properties (in this case type) need to be displayed with different labels depending on the context. For example the property rdf:type (= @type) should be displayed as just 'Type' for works whereas for instances it should be 'Carrier type'. For this reason another entry with the key alias will now be included in a Slice or SearchMapping when needed to point out which label should be used (this feature is added in libris/librisxl#1416).