Block id.kb.se and libris.kb.se datasets from being edited (#1082)
Do a simple hardcoded check for dataset URI prefixes for now.
Until we have decided on an how to specify this in backend.
See for example EntityContainers for Concepts.

Prevents accidental editing of language, library and other definitions from externally defined datasets.
There is still no access control mechanism for these in backend. This does not guard against malicious changes.