Fix resource image placement on Safari (#1084)
Solves weird placement of resource images in Safari.

Clearly, Safari doesn't respect the max-height of an element with display:table (introduced with the figcaption stuff). Using only height works - it doesn't seem to break anything else...

Setting w-full on the image container allows the image to align to the page left margin on md+ screens. This seems to be desired since it has justify-center md:justify-start.