Fix display of user's own holding in hitlist (#1087)
Was broken by the early return in c4434a0.
We can always find the holdings in `@reverse/itemOf` for each Instance
So remove the code that does an additional query for the holding.

TODO
A further improvement would be to display the holding link directly without the spinner when viewing the record in the inspector. Since we don't need to wait for the query anymore.