fix(supersearch): Remove redundant text-selection code
If the expanded search text box is placed at the same place as the collapsed, text-selection by double-clicking and triple-clicking will work mostly natively out-of-the-box.

If we decide that we need to support cases where the expanded search text box is placed somewhere else then https://github.com/libris/lxlviewer/pull/1171 can be merged.