fix(cataloging): Some records are slow to copy
When copying a record, any missing references are fetched and added to
quoted. Copying a record with many links was slow. Because each thing
added to quoted would be a commit on the store and also clone the whole
quouted commit. Instead make it possible to pass an array of things to
`addToQuoted` which will be commited at once.