feat(cataloging): Display backend validation errors
Quick and dirty display of JSON-LD validation errors from backend.
These might happen on old records that are broken or on when creating
a new record from a user uploaded file.

When we get a HTTP 400 because of invalid JSON-LD inside the record we
also get a JSON structure describing the individual errors. Display
these inside the form.

Also display a more specific error toast that the generic "Something
went wrong".

Added swedish translations which are hopefully a bit more helpful for users.
(We should probably use the error code/type (e.g. VOCAB_STRING_EXPECTED)
that is also present in the error object as key for i18n but it doesn't
fit with the way i18n.json is set up right now.)