feat(lxlweb): citations (LWS-94) (#1359)
LWS-94

Solves
======

A barebones implementation of citation features for LWS. More to be added (see TODO), but seems like it's time for a review. It contains:

Zotero support
--------------
Using the same implementation as legacy Libris, with a script tag pointing to an unapi endpoint and <abbr> tags including the resource id. This is enough for Zotero to be able to consume metadata.

The api follows the unapi v1 spec, is located at http://localhost:5173/api/sv/cite and serves CSL, bibtex and RIS
For reference, the old Libris api is located here

Internally, the api uses the citations.ts utils 👇

Citation.js + helpers
---------------------
The citation.ts util file contains various helpers, including a wrapper around the citation.js library, import of the static assets (citation style files etc). This is a large dependency, but felt necessary because of its plugin ecosystem that can produce various outputs including bibtex and RIS. The CSL plugin uses citeproc-js internally, meaning we can use the official CSL collection of styles and locales.

This file also includes a very rudimentary mapping of our data to CSL. A couple of examples of its current shortcomings:
* type hardcoded to "book"
* author simply picks the first contributor

On the fly-citation on the frontend
-----------------------------------
You can currently cite only a single instance in the UI, using a link in the resource heading or in the instances list.
the cite=id opens a CitationsModal and Citations.svelte produces the formatting on demand using the utils ☝️.

In theory, we could reuse the fetched resource data here. But the csl mapping relies on computedLabel (for titles etc) that the resource data does not have.

So, another fetch is needed. But to avoid triggering the massive load function again, a separate layout.server was created. the layout.server load now only reacts to cite param change, fetches the resource in CSL format using the cite api and passes it as page.data.citations.

TODO
====
* Proper mapping KBV -> CSL
* Improve asset storage/reduce bundle size??

Summary of changes
==================
* Add package citation.js + create citation.ts util functions
* Add CitationsModal and Citations components
* Refactor [fnurgel]/page.server and [fnurgel]/layout.server files
* Add api/[lang]/cite API route