feat(lxlweb/data): Improve Classification display (#1436)
Improve rendering of Classification.

    Show inScheme / type after code so that classification schemes can be distinguished.
    / before ConceptScheme version
    ; between codes in the rare case where there are multiple inside one classification object.
    Show additionalClassificationDdc property

Make it possible to use rdf:type inside showProperties. Instead of the old way of using displayType() in resourceStyle. This makes it possible to place it anywhere inside the properties, style it and have label translation working. e.g. "DDK-klassifikation below". (they should maybe get a shorter label in the future "DDK")