fix(lxlweb): don't add wildcard, use _mappingOnly (#1440)
Solves

Hopefully fixes a problem 1) and paves the way to fixing another 2)

    The wildcard added by frontend for empty qualifier values resulted in queries like contribution:(*) - in turn causing endless spinners and ES timing out.

    Basically "Sökning som krockar med frågespråket ska inte generera fel". If this is to be solved in the frontend, qualifierKeys must be validated quickly by the api.

The solution here is to refactor insertWildcard to a helper that returns searchParams for different cases.

In the case of an empty qualifierValue (contribution:()), we replace it with contribution:"" (as long as empty values are disallowed) and use _mappingOnly - i.e no suggestions until you start typing

In the case of a qualifierValue (contribution:(astrid)), we instead apply _suggest.

So the noticeable difference here is no wildcard suggestions, fast labels.
Summary of changes

    insertWildcard -> addSuggestParams
    Update tests